requirements.txt
setuptools
wheel
wit
gTTS
pyowm
geocoder
SpeechRecognition
youtube_dl
pafy
python-vlc
beautifulsoup4
